feat(ImageLocation):some minor refactors
